+++ R/simulation-simInit.R
@@ -222,7 +222,7 @@ utils::globalVariables(c(".", "Package", "hasVersionSpec"))
 #'    .globals = list(stackName = "landscape", burnStats = "nPixelsBurned")
 #'  ),
 #'  modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
-#'  paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
+#'  paths = list(modulePath = getSampleModules(tempdir()))
 #' )
 #' spades(mySim) # shows plotting
 #'
@@ -234,7 +234,7 @@ utils::globalVariables(c(".", "Package", "hasVersionSpec"))
 #'    fireSpread = list(.plotInitialTime = NA)
 #'  ),
 #'  modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
-#'  paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
+#'  paths = list(modulePath = getSampleModules(tempdir()))
 #' )
 #' outSim <- spades(mySim)
 #'
@@ -246,7 +246,7 @@ utils::globalVariables(c(".", "Package", "hasVersionSpec"))
 #'     .globals = list(stackName = "landscape", burnStats = "nPixelsBurned")
 #'   ),
 #'   modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
-#'   paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"),
+#'   paths = list(modulePath = getSampleModules(tempdir()),
 #'                outputPath = tempdir()),
 #'   inputs = data.frame(
 #'     files = dir(file.path(mapPath), full.names = TRUE, pattern = "tif")[1:2],
@@ -268,7 +268,7 @@ utils::globalVariables(c(".", "Package", "hasVersionSpec"))
 #'      randomLandscapes = list(nx = 10, ny = 10)
 #'      ),
 #'    paths = list(
-#'      modulePath = system.file("sampleModules", package = "SpaDES.core"),
+#'      modulePath = getSampleModules(tempdir()),
 #'      outputPath = tempdir()
 #'    )
 #'  )
@@ -293,7 +293,7 @@ utils::globalVariables(c(".", "Package", "hasVersionSpec"))
 #'      .globals = list(stackName = "landscape", burnStats = "nPixelsBurned")
 #'    ),
 #'    modules = list("randomLandscapes", "fireSpread", "caribouMovement"),
-#'    paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"),
+#'    paths = list(modulePath = getSampleModules(tempdir()),
 #'                 outputPath = tempdir()),
 #'    inputs = data.frame(
 #'      files = dir(file.path(mapPath), full.names = TRUE, pattern = "tif")[1:2],