+++ inst/examples/example_moduleMetadata.R
@@ -1,9 +1,9 @@
-path <- system.file("sampleModules", package = "SpaDES.core")
-sampleModules <- dir(path)
 ## turn off code checking -- don't need it here
 opts <- options("spades.moduleCodeChecks" = FALSE,
                 "spades.useRequire" = FALSE)
 
+path <- getSampleModules(tempdir())
+sampleModules <- dir(path)
 x <- moduleMetadata(sampleModules[3], path = path)
 
 ## using simList
@@ -14,7 +14,7 @@ if (require("SpaDES.tools", quietly = TRUE)) {
         .globals = list(stackName = "landscape")
       ),
       modules = list("caribouMovement"),
-      paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
+      paths = list(modulePath = path)
    )
    moduleMetadata(sim = mySim)
 }