+++ inst/examples/example_moduleVersion.R
@@ -1,5 +1,5 @@
 # using filepath
-path <- system.file("sampleModules", package = "SpaDES.core")
+path <- getSampleModules(tempdir())
 moduleVersion("caribouMovement", path)
 
 # using simList